# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  request.AppKnowledgeBase:
    properties:
      id:
        description: 知识库id
        type: string
      metaDataFilterParams:
        $ref: '#/definitions/request.MetaDataFilterParams'
      name:
        type: string
    required:
    - id
    type: object
...
  request.CreateKnowledgeReq:
    properties:
      description:
        type: string
      embeddingModelInfo:
        $ref: '#/definitions/request.EmbeddingModel'
      name:
        type: string
    required:
    - embeddingModelInfo
    - name
    type: object
...
  request.DeleteKnowledge:
    properties:
      knowledgeId:
        type: string
    required:
    - knowledgeId
    type: object
...
  request.DocMetaData:
    properties:
      metaId:
        description: 元数据id
        type: string
      metaKey:
        description: key
        type: string
      metaRule:
        description: 正则表达式
        type: string
      metaValue:
        description: 确定值
        type: string
      metaValueType:
        description: string，number，time
        type: string
      option:
        description: option:add(新增)、update(更新)、delete(删除),update 和delete 的时候metaId
          不能为空
        type: string
    type: object
...
  request.EmbeddingModel:
    properties:
      modelId:
        type: string
    required:
    - modelId
    type: object
...
  request.KnowledgeHitReq:
    properties:
      knowledgeList:
        items:
          $ref: '#/definitions/request.AppKnowledgeBase'
        type: array
      knowledgeMatchParams:
        $ref: '#/definitions/request.KnowledgeMatchParams'
      question:
        type: string
    required:
    - knowledgeMatchParams
    - question
    type: object
  request.KnowledgeMatchParams:
    properties:
      keywordPriority:
        description: 关键词权重
        type: number
      matchType:
        description: matchType：vector（向量检索）、text（文本检索）、mix（混合检索：向量+文本）
        type: string
      priorityMatch:
        description: 权重匹配，只有在混合检索模式下，选择权重设置后，这个才设置为1
        type: integer
      rerankModelId:
        description: rerank模型id
        type: string
      semanticsPriority:
        description: 语义权重
        type: number
      termWeight:
        description: 关键词系数
        type: number
      termWeightEnable:
        description: 关键词系数开关
        type: boolean
      threshold:
        description: threshold 过滤分数阈值
        type: number
      topK:
        description: topK 获取最高的几行
        type: integer
    required:
    - matchType
    type: object
  request.KnowledgeMetaValueListReq:
    properties:
      docIdList:
        items:
          type: string
        type: array
      knowledgeId:
        type: string
    required:
    - docIdList
    - knowledgeId
    type: object
  request.KnowledgeSelectReq:
    properties:
      name:
        type: string
      tagId:
        items:
          type: string
        type: array
    type: object
...
  request.MetaDataFilterParams:
    properties:
      filterEnable:
        description: 元数据过滤开关
        type: boolean
      filterLogicType:
        description: 元数据逻辑条件：and/or
        type: string
      metaFilterParams:
        description: 元数据过滤参数列表
        items:
          $ref: '#/definitions/request.MetaFilterParams'
        type: array
    type: object
  request.MetaFilterParams:
    properties:
      condition:
        description: 条件
        type: string
      key:
        description: Key
        type: string
      type:
        description: 类型（Time, String, Number）
        type: string
      value:
        description: value
        type: string
    type: object
...
  request.UpdateKnowledgeReq:
    properties:
      description:
        type: string
      knowledgeId:
        type: string
      name:
        type: string
    required:
    - knowledgeId
    - name
    type: object
...
  request.UpdateMetaValueReq:
    properties:
      applyToSelected:
        type: boolean
      docIdList:
        items:
          type: string
        type: array
      knowledgeId:
        type: string
      metaValueList:
        items:
          $ref: '#/definitions/request.DocMetaData'
        type: array
    required:
    - docIdList
    - knowledgeId
    type: object
...
  response.ChildContent:
    properties:
      childSnippet:
        type: string
      score:
        type: number
    type: object
...
  response.ChunkSearchList:
    properties:
      childContentList:
        items:
          $ref: '#/definitions/response.ChildContent'
        type: array
      childScore:
        items:
          type: number
        type: array
      knowledgeName:
        type: string
      snippet:
        type: string
      title:
        type: string
    type: object
...
  response.CreateKnowledgeResp:
    properties:
      knowledgeId:
        type: string
    type: object
...
  response.EmbeddingModelInfo:
    properties:
      modelId:
        type: string
    type: object
...
  response.GetKnowledgeMetaSelectResp:
    properties:
      knowledgeMetaList:
        items:
          $ref: '#/definitions/response.KnowledgeMetaItem'
        type: array
    type: object
...
  response.KnowledgeHitResp:
    properties:
      prompt:
        description: 提示词列表
        type: string
      score:
        description: 打分信息
        items:
          type: number
        type: array
      searchList:
        description: 种种结果
        items:
          $ref: '#/definitions/response.ChunkSearchList'
        type: array
    type: object
  response.KnowledgeInfo:
    properties:
      createAt:
        description: 创建时间
        type: string
      createUserId:
        type: string
      description:
        description: 知识库描述
        type: string
      docCount:
        description: 文档数量
        type: integer
      embeddingModelInfo:
        allOf:
        - $ref: '#/definitions/response.EmbeddingModelInfo'
        description: embedding模型信息
      knowledgeId:
        description: 知识库id
        type: string
      knowledgeTagList:
        description: 知识库标签列表
        items:
          $ref: '#/definitions/response.KnowledgeTag'
        type: array
      name:
        description: 知识库名称
        type: string
      orgName:
        description: 知识库所属名称
        type: string
      permissionType:
        description: '权限类型:0: 查看权限; 10: 编辑权限; 20: 授权权限,数值不连续的原因防止后续有中间权限，目前逻辑 授权权限>编辑权限>查看权限'
        type: integer
      share:
        description: 是分享，还是私有
        type: boolean
    type: object
  response.KnowledgeListResp:
    properties:
      knowledgeList:
        items:
          $ref: '#/definitions/response.KnowledgeInfo'
        type: array
    type: object
  response.KnowledgeMetaItem:
    properties:
      metaId:
        type: string
      metaKey:
        type: string
      metaValue:
        description: 确定值
        type: string
      metaValueType:
        type: string
    type: object
  response.KnowledgeMetaValueListResp:
    properties:
      knowledgeMetaValues:
        items:
          $ref: '#/definitions/response.KnowledgeMetaValues'
        type: array
    type: object
  response.KnowledgeMetaValues:
    properties:
      metaId:
        type: string
      metaKey:
        type: string
      metaValue:
        description: 确定值
        items:
          type: string
        type: array
      metaValueType:
        type: string
    type: object
...
  response.KnowledgeTag:
    properties:
      selected:
        description: 此表标签是否选中
        type: boolean
      tagId:
        description: 知识库标签id
        type: string
      tagName:
        description: 知识库标签名称
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    ## HTTP Header
    | Header        | 说明      |
    | ------------- | --------- |
    | Authorization | JWT token |
    | X-Language    | 语言Code  |
    | X-Org-Id      | 组织ID    |
    | X-Client-Id   | 客户端标识|

    ## HTTP Status
    | HTTP Status             | 说明                   |
    | ----------------------- | ---------------------- |
    | 200, StatusOK           | 请求返回成功           |
    | 400, StatusBadRequest   | 请求返回失败，用于业务 |
    | 401, StatusUnauthorized | JWT认证失败            |
    | 403, StatusForbidden    | 没有权限               |

    ## 权限-菜单对应表
    | 一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |
    |-------------|-------|------|------|------|------|
    | guest       |       |      | 【访客】 |      |      |
    | common      |       |      | 【通用】 |      |      |
    | permission  |       |      | 权限管理 |      |      |
    | permission  | user  |      | 权限管理 | 用户管理 |      |
    | permission  | org   |      | 权限管理 | 组织管理 |      |
    | permission  | role  |      | 权限管理 | 角色管理 |      |

    ## `/v1/user/permission`返回用例
    ```json
    {
      "code": 0,
      "data": {
        "orgPermission": {
          "org": {"id": "test-org-id", "name": "test-org-name"},
          "permissions": [
            {"perm": "permission"},
            {"perm": "permission.user"},
            {"perm": "permission.org"},
            {"perm": "permission.role"}
          ]
        }
      },
      "msg": "操作成功"
    }
    ```
...
paths:
...
  /knowledge:
    delete:
      consumes:
      - application/json
      description: 删除知识库
      parameters:
      - description: 删除知识库请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteKnowledge'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除知识库
      tags:
      - knowledge
    post:
      consumes:
      - application/json
      description: 创建知识库
      parameters:
      - description: 创建知识库请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateKnowledgeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.CreateKnowledgeResp'
                  type: array
              type: object
      security:
      - JWT: []
      summary: 创建知识库
      tags:
      - knowledge
    put:
      consumes:
      - application/json
      description: 修改知识库（文档分类）
      parameters:
      - description: 修改知识库请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateKnowledgeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 修改知识库（文档分类）
      tags:
      - knowledge
...
  /knowledge/hit:
    post:
      consumes:
      - application/json
      description: 知识库命中测试
      parameters:
      - description: 知识库命中测试请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeHitReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowledgeHitResp'
              type: object
      security:
      - JWT: []
      summary: 知识库命中测试
      tags:
      - knowledge
...
  /knowledge/meta/select:
    get:
      consumes:
      - application/json
      description: 获取知识库元数据
      parameters:
      - in: query
        name: knowledgeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.GetKnowledgeMetaSelectResp'
              type: object
      security:
      - JWT: []
      summary: 获取知识库元数据
      tags:
      - knowledge
  /knowledge/meta/value/list:
    post:
      consumes:
      - application/json
      description: 获取文档元数据列表
      parameters:
      - description: 文档元数据列表请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeMetaValueListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowledgeMetaValueListResp'
              type: object
      security:
      - JWT: []
      summary: 获取文档元数据列表
      tags:
      - knowledge
  /knowledge/meta/value/update:
    post:
      consumes:
      - application/json
      description: 更新知识库元数据值
      parameters:
      - description: 更新知识库元数据值请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateMetaValueReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新知识库元数据值
      tags:
      - knowledge
...
  /knowledge/select:
    post:
      consumes:
      - application/json
      description: 查询知识库列表
      parameters:
      - description: 查询知识库列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeSelectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowledgeListResp'
              type: object
      security:
      - JWT: []
      summary: 查询知识库列表
      tags:
      - knowledge
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"